Fixing out names are written.
A. I think the line sb.append("\",\""); would print the wrong thing.
(It'd print '","' right? Don't we want '"",'?)
B. It's not necessary (right?). We already did a null-check above. We
don't care about fdd anymore. The names may be empty strings, but that's
OK too.

(Also the fact that I searched my cardinal export output for "," and
found zero matches suggests fdd is never null, right? I don't want to
remove the null-check above, though, because if neither of us are sure
it won't hurt...)