Scaling up logging for debugging
This fixes at least one vital bug: the doCreateQueryIterator had one
early point that returned a QueryIterator instead of a
QueryIteratorDash. We need the QueryIteratoDash to listen to the
iterator and update our stats, otherwise we never graduate to the other
optimizations. (There's also a SEVERE level logging event now that
alerts us if that happens again...)