Migrating xray code to birdcage project from pump project
Originally I put it in the pump project because the pump project was the
only repo I maintained, and I wanted to clearly distinguish that this
code was something I developed in my spare time that I owned.

But now I also maintain the birdcage project, and it's a better fit
here. (It's only known use is in the birdcage project.)