Supporting both MST and MTC approaches
I don't have Aspen sites to test against, and I'm afraid (with no way to
verify) that we may run into cases where the MST data is incomplete or
missing, and the MTC data still contains a useful reference to the
primary teacher.

So out of an abundance of caution: let's poll both. Redundant OneRoster
beans are dropped when we save.