Adding output to flag unresolved grade levels
Jen is having problems with kindergarten students not having the
appropriate grade level coming through. On her end it looks like their
code is "KF", which should be something we allow for. (It's in the
GradeLevel code -- which is part of a separate library.)

So I don't understand this problem yet; hopefully this additional
feedback will help explain it.