Adding getGradeLevel(..)
Surprisingly Jen reported that "KF" was not being identifier as
GradeLevel.KG. But the GradeLevel.java source code appears to check for
"KF"? It's possible the source code I have is out-of-date. (It's based
on a snapshot of the customizations.jar that included in the source.) Or
maybe Jen has an older jar file? Not sure. Either way: this might be
sufficient.